Refactoring and minor changes
