--- src/test/java/usecases/DummyLoadInteractor.java
@@ -1,4 +1,0 @@
-package usecases;
-
-public class DummyLoadInteractor {
-}
