+++ src/test/java/usecases/LoginInteractorTest.java
@@ -2,6 +2,7 @@ package usecases;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import usecases.dummyclasses.DummyLoginInteractor;
 
 
 import java.util.HashMap;
