+++ src/test/java/usecases/dummyclasses/DummyLoginInteractor.java
@@ -1,4 +1,4 @@
-package usecases;
+package usecases.dummyclasses;
 
 import java.io.*;
 import java.util.HashMap;
